+++ pennylane/plugins/default_qubit.py
@@ -35,7 +35,7 @@ Auxiliary functions
 -------------------
 
 .. autosummary::
-    spectral_decomposition_qubit
+    spectral_decomposition
     unitary
     hermitian
 
@@ -55,6 +55,10 @@ Gates and operations
     CNOT
     SWAP
     CZ
+    CRotx
+    CRoty
+    CRotz
+    CRot3
 
 Observables
 ------------
@@ -73,10 +77,12 @@ Classes
 Code details
 ^^^^^^^^^^^^
 """
+from collections import OrderedDict
+import itertools
 import warnings
 
 import numpy as np
-from scipy.linalg import expm, eigh
+from scipy.linalg import eigh
 
 from pennylane import Device
 
@@ -89,19 +95,19 @@ tolerance = 1e-10
 #  utilities
 #========================================================
 
-def spectral_decomposition_qubit(A):
-    r"""Spectral decomposition of a :math:`2\times 2` Hermitian matrix.
+def spectral_decomposition(A):
+    r"""Spectral decomposition of a Hermitian matrix.
 
     Args:
-        A (array): :math:`2\times 2` Hermitian matrix
+        A (array): Hermitian matrix
 
     Returns:
         (vector[float], list[array[complex]]): (a, P): eigenvalues and hermitian projectors
-        such that :math:`A = \sum_k a_k P_k`.
+            such that :math:`A = \sum_k a_k P_k`.
     """
     d, v = eigh(A)
     P = []
-    for k in range(2):
+    for k in range(d.shape[0]):
         temp = v[:, k]
         P.append(np.outer(temp, temp.conj()))
     return d, P
@@ -147,7 +153,7 @@ def Rotx(theta):
     Returns:
         array: unitary 2x2 rotation matrix :math:`e^{-i \sigma_x \theta/2}`
     """
-    return expm(-1j * theta/2 * X)
+    return np.cos(theta/2) * I + 1j * np.sin(-theta/2) * X
 
 
 def Roty(theta):
@@ -158,7 +164,7 @@ def Roty(theta):
     Returns:
         array: unitary 2x2 rotation matrix :math:`e^{-i \sigma_y \theta/2}`
     """
-    return expm(-1j * theta/2 * Y)
+    return np.cos(theta/2) * I + 1j * np.sin(-theta/2) * Y
 
 
 def Rotz(theta):
@@ -169,7 +175,7 @@ def Rotz(theta):
     Returns:
         array: unitary 2x2 rotation matrix :math:`e^{-i \sigma_z \theta/2}`
     """
-    return expm(-1j * theta/2 * Z)
+    return np.cos(theta/2) * I + 1j * np.sin(-theta/2) * Z
 
 
 def Rot3(a, b, c):
@@ -181,6 +187,51 @@ def Rot3(a, b, c):
         array: unitary 2x2 rotation matrix ``rz(c) @ ry(b) @ rz(a)``
     """
     return Rotz(c) @ (Roty(b) @ Rotz(a))
+
+
+def CRotx(theta):
+    r"""Two-qubit controlled rotation about the x axis.
+
+    Args:
+        theta (float): rotation angle
+    Returns:
+        array: unitary 4x4 rotation matrix :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R_x(\theta)`
+    """
+    return np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, np.cos(theta/2), -1j*np.sin(theta/2)], [0, 0, -1j*np.sin(theta/2), np.cos(theta/2)]])
+
+
+def CRoty(theta):
+    r"""Two-qubit controlled rotation about the y axis.
+
+    Args:
+        theta (float): rotation angle
+    Returns:
+        array: unitary 4x4 rotation matrix :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R_y(\theta)`
+    """
+    return np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, np.cos(theta/2), -np.sin(theta/2)], [0, 0, np.sin(theta/2), np.cos(theta/2)]])
+
+
+def CRotz(theta):
+    r"""Two-qubit controlled rotation about the z axis.
+
+    Args:
+        theta (float): rotation angle
+    Returns:
+        array: unitary 4x4 rotation matrix :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R_z(\theta)`
+    """
+    return np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, np.exp(-1j*theta/2), 0], [0, 0, 0, np.exp(1j*theta/2)]])
+
+
+def CRot3(a, b, c):
+    r"""Arbitrary two-qubit controlled rotation using three Euler angles.
+
+    Args:
+        a,b,c (float): rotation angles
+    Returns:
+        array: unitary 4x4 rotation matrix :math:`|0\rangle\langle 0|\otimes \mathbb{I}+|1\rangle\langle 1|\otimes R(a,b,c)`
+    """
+    return np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, np.exp(-1j*(a+c)/2)*np.cos(b/2), -np.exp(1j*(a-c)/2)*np.sin(b/2)], [0, 0, np.exp(-1j*(a-c)/2)*np.sin(b/2), np.exp(1j*(a+c)/2)*np.cos(b/2)]])
+
 
 
 #========================================================
@@ -249,7 +300,7 @@ class DefaultQubit(Device):
     """
     name = 'Default qubit PennyLane plugin'
     short_name = 'default.qubit'
-    pennylane_requires = '0.4'
+    pennylane_requires = '0.5'
     version = '0.4.0'
     author = 'Xanadu Inc.'
 
@@ -271,7 +322,11 @@ class DefaultQubit(Device):
         'RX': Rotx,
         'RY': Roty,
         'RZ': Rotz,
-        'Rot': Rot3
+        'Rot': Rot3,
+        'CRX': CRotx,
+        'CRY': CRoty,
+        'CRZ': CRotz,
+        'CRot': CRot3
     }
 
     _observable_map = {
@@ -298,6 +353,8 @@ class DefaultQubit(Device):
                 self._state = state
             else:
                 raise ValueError('State vector must be of length 2**wires.')
+            if wires is not None and wires != [] and list(wires) != list(range(self.num_wires)):
+                raise ValueError("The default.qubit plugin can apply QubitStateVector only to all of the {} wires.".format(self.num_wires))
             return
         if operation == 'BasisState':
             n = len(par[0])
@@ -345,23 +402,44 @@ class DefaultQubit(Device):
         return np.reshape(state_multi_index, 2 ** self.num_wires)
 
     def expval(self, observable, wires, par):
-        A = self._get_operator_matrix(observable, par)
         if self.shots == 0:
             # exact expectation value
+            A = self._get_operator_matrix(observable, par)
             ev = self.ev(A, wires)
         else:
             # estimate the ev
-            # sample Bernoulli distribution n_eval times / binomial distribution once
-            a, P = spectral_decomposition_qubit(A)
-            p0 = self.ev(P[0], wires)  # probability of measuring a[0]
-            n0 = np.random.binomial(self.shots, p0)
-            ev = (n0*a[0] +(self.shots-n0)*a[1]) / self.shots
+            ev = np.mean(self.sample(observable, wires, par, self.shots))
 
         return ev
 
     def var(self, observable, wires, par):
+        if self.shots == 0:
+            # exact expectation value
+            A = self._get_operator_matrix(observable, par)
+            var = self.ev(A@A, wires) - self.ev(A, wires)**2
+        else:
+            # estimate the ev
+            var = np.var(self.sample(observable, wires, par, self.shots))
+
+        return var
+
+    def sample(self, observable, wires, par, n=None):
+        if n is None:
+            n = self.shots
+
+        if n == 0:
+            raise ValueError("Calling sample with n = 0 is not possible.")
+        if n < 0 or not isinstance(n, int):
+            raise ValueError("The number of samples must be a positive integer.")
+
         A = self._get_operator_matrix(observable, par)
-        return self.ev(A@A, wires) - self.ev(A, wires)**2
+        a, P = spectral_decomposition(A)
+
+        p = np.zeros(a.shape)
+        for idx, Pi in enumerate(P):
+            p[idx] = self.ev(Pi, wires)
+
+        return np.random.choice(a, n, p=p)
 
     def _get_operator_matrix(self, operation, par):
         """Get the operator matrix for a given operation or observable.
@@ -378,6 +456,14 @@ class DefaultQubit(Device):
         return A(*par)
 
     def ev(self, A, wires):
+        r"""Expectation value of observable on specified wires.
+
+         Args:
+          A (array[float]): the observable matrix as array
+          wires (Sequence[int]): target subsystems
+         Returns:
+          float: expectation value :math:`\expect{A} = \bra{\psi}A\ket{\psi}`
+        """
         As = self.mat_vec_product(A, self._state, wires)
         expectation = np.vdot(self._state, As)
 
@@ -398,3 +484,12 @@ class DefaultQubit(Device):
     @property
     def observables(self):
         return set(self._observable_map.keys())
+
+    def probability(self):
+        if self._state is None:
+            return None
+
+        states = itertools.product(range(2), repeat=self.num_wires)
+        probs = np.abs(self._state)**2
+
+        return OrderedDict(zip(states, probs))