+++ tests/test_templates_layers.py
@@ -40,7 +40,7 @@ class TestInterferometer:
         @qml.qnode(dev)
         def circuit(varphi, mesh):
             Interferometer(theta=None, phi=None, varphi=varphi, mesh=mesh, wires=0)
-            return qml.expval(qml.MeanPhoton(0))
+            return qml.expval(qml.NumberOperator(0))
 
         with pytest.raises(QuantumFunctionError) as excinfo:
             circuit(varphi, 'rectangular')
@@ -50,7 +50,7 @@ class TestInterferometer:
         @qml.qnode(dev)
         def circuit(varphi, bs):
             Interferometer(theta=None, phi=None, varphi=varphi, beamsplitter=bs, wires=0)
-            return qml.expval(qml.MeanPhoton(0))
+            return qml.expval(qml.NumberOperator(0))
 
         with pytest.raises(QuantumFunctionError) as excinfo:
             circuit(varphi, 'clements')
@@ -69,11 +69,11 @@ class TestInterferometer:
 
         def circuit_rect(varphi):
             Interferometer(theta, phi, varphi, mesh='rectangular', beamsplitter='clements', wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         def circuit_tria(varphi):
             Interferometer(theta, phi, varphi, mesh='triangular', beamsplitter='clements', wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         for c in [circuit_rect, circuit_tria]:
             qnode = qml.QNode(c, dev)
@@ -98,7 +98,7 @@ class TestInterferometer:
 
         def circuit(varphi):
             Interferometer(theta=None, phi=None, varphi=varphi, wires=0)
-            return qml.expval(qml.MeanPhoton(0))
+            return qml.expval(qml.NumberOperator(0))
 
         qnode = qml.QNode(circuit, dev)
         assert np.allclose(qnode(varphi), 0, atol=tol)
@@ -121,7 +121,7 @@ class TestInterferometer:
 
         def circuit(varphi):
             Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         qnode = qml.QNode(circuit, dev)
         assert np.allclose(qnode(varphi), [0, 0], atol=tol)
@@ -148,7 +148,7 @@ class TestInterferometer:
 
         def circuit(varphi):
             Interferometer(theta, phi, varphi, mesh='triangular', wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         qnode = qml.QNode(circuit, dev)
         assert np.allclose(qnode(varphi), [0, 0], atol=tol)
@@ -175,7 +175,7 @@ class TestInterferometer:
 
         def circuit(varphi):
             Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         qnode = qml.QNode(circuit, dev)
         assert np.allclose(qnode(varphi), [0, 0], atol=tol)
@@ -204,11 +204,11 @@ class TestInterferometer:
 
         def circuit_rect(varphi):
             Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         def circuit_tria(varphi):
             Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         for c in [circuit_rect, circuit_tria]:
             # test both meshes (both give identical results for the 3 mode case).
@@ -242,7 +242,7 @@ class TestInterferometer:
 
         def circuit_rect(varphi):
             Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         qnode = qml.QNode(circuit_rect, dev)
         assert np.allclose(qnode(varphi), [0]*N, atol=tol)
@@ -274,7 +274,7 @@ class TestInterferometer:
 
         def circuit_tria(varphi):
             Interferometer(theta, phi, varphi, mesh='triangular', wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         qnode = qml.QNode(circuit_tria, dev)
         assert np.allclose(qnode(varphi), [0]*N, atol=tol)
@@ -315,7 +315,7 @@ class TestInterferometer:
                 qml.Squeezing(sq[w][0], sq[w][1], wires=w)
 
             Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.MeanPhoton(w)) for w in wires]
+            return [qml.expval(qml.NumberOperator(w)) for w in wires]
 
         res = circuit(theta, phi, varphi)
         expected = np.array([0.96852694, 0.23878521, 0.82310606, 0.16547786])
@@ -548,11 +548,10 @@ class TestRandomLayers:
     def rots(self, request):
         return request.param
 
-    def test_random_layers_seed(self, n_layers, tol, seed):
-        """Test that pennylane.templates.layers.RandomLayers() gets deterministic when using seed."""
+    def test_random_layers_seed_deterministic(self, n_layers, tol, seed):
+        """Test that pennylane.templates.layers.RandomLayers() acts deterministically when using fixed seed."""
         n_rots = 1
         n_wires = 2
-        impr = CNOT
         dev = qml.device('default.qubit', wires=n_wires)
         weights = np.random.randn(n_layers, n_rots)
 
@@ -568,6 +567,46 @@ class TestRandomLayers:
         qnode2 = qml.QNode(circuit2, dev)
 
         assert np.allclose(qnode1(weights), qnode2(weights), atol=tol)
+
+    def test_random_layers_no_seed_deterministic(self, n_layers, tol):
+        """Test that pennylane.templates.layers.RandomLayers() acts deterministically when using no seed."""
+        n_rots = 1
+        n_wires = 2
+        dev = qml.device('default.qubit', wires=n_wires)
+        weights = np.random.randn(n_layers, n_rots)
+
+        def circuit1(weights):
+            RandomLayers(weights=weights, wires=range(n_wires))
+            return qml.expval(qml.PauliZ(0))
+
+        def circuit2(weights):
+            RandomLayers(weights=weights, wires=range(n_wires))
+            return qml.expval(qml.PauliZ(0))
+
+        qnode1 = qml.QNode(circuit1, dev)
+        qnode2 = qml.QNode(circuit2, dev)
+
+        assert np.allclose(qnode1(weights), qnode2(weights), atol=tol)
+
+    def test_random_layers_two_seeds_different(self, n_layers, tol):
+        """Test that pennylane.templates.layers.RandomLayers() does not have the same output for two different seeds."""
+        n_rots = 10
+        n_wires = 2
+        dev = qml.device('default.qubit', wires=n_wires)
+        weights = np.random.randn(n_layers, n_rots)
+
+        def circuit1(weights):
+            RandomLayers(weights=weights, wires=range(n_wires), seed=0)
+            return qml.expval(qml.PauliZ(0))
+
+        def circuit2(weights):
+            RandomLayers(weights=weights, wires=range(n_wires), seed=1)
+            return qml.expval(qml.PauliZ(0))
+
+        qnode1 = qml.QNode(circuit1, dev)
+        qnode2 = qml.QNode(circuit2, dev)
+
+        assert not np.allclose(qnode1(weights), qnode2(weights), atol=tol)
 
     def test_random_layers_nlayers(self, n_layers):
         """Test that  pennylane.templates.layers.RandomLayers() picks the correct number of gates."""